requirements.txt
# Core Flask dependencies
flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Code Q&A dependencies
lancedb>=0.4.0
openai>=1.0.0
tree-sitter>=0.21.3
pandas>=2.0.0
markdown>=3.4.0

# GDD RAG dependencies
dashscope>=1.25.0
# NOTE: lightrag and lightrag-hku require Python 3.10+ (uses 'match' statement)
# These are only needed for legacy RAGAnything integration, not for core Supabase-based functionality
# If you need RAGAnything features, upgrade to Python 3.10+ and uncomment:
# lightrag>=0.1.0b6
# lightrag-hku>=1.4.9.8
docling>=1.0.0  # PDF to Markdown conversion (optional - falls back to PyPDF2 if not installed)

# Keyword Extractor backenpendencies (for document upload/processing)
langchain-text-splitters>=0.0.1  # Text chunking utilities
pypdf2>=3.0.0  # PDF fallback extraction

# Supabase (for storage)
supabase>=2.0.0
pgvector>=0.2.0

# Optional but recommended
redis>=5.0.0
cohere>=4.0.0

# Translation and Language Detection
deep-translator>=1.11.4  # Free Google Translate API (no API key needed)
fast-langdetect>=0.1.0  # Fast language detection (80x faster than fastText)
nltk>=3.8.1  # Natural Language Toolkit (for WordNet synonym generation)
# Note: NLTK data (wordnet, omw-1.4) is downloaded automatically on first use
